refactor: modularize landing page sections
