+++ composables/useWhatsApp.ts
@@ -1,0 +1,4 @@
+export const useWhatsApp = () => {
+  const buildLink = (phone: string, message: string) => `https://wa.me/${phone}?text=${encodeURIComponent(message)}`;
+  return { buildLink };
+};
